qcacld-3.0: change HAL msg type macros to enum
The HAL msg type that were defined as macros are converted
to enum halmsgtype. This is done as it is difficult to parse
macros directly in any post processing script.
Since, driver maintains a global list of messages (qdf trace)
which different modules posts to each other. For processing
these list of messages in a common post processing script,
it is required to convert the macros to enum as enum are
easy to process in the script.

Change-Id: I3938ff02a54a2fca9e831ea1f8d9389e8b54df01
CRs-Fixed: 2719659
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>