qcacld-3.0: Fix to have single PMK cache entry in FT roam
In FT roam when the CSR lookup fails then the PMK details from
the roam sync indication will be updated to Session/PMK cache.
This will result in having multiple PMK cache entries for the
same MDID.
This fix is to avoid updating the Session/PMK cache in case of
FT roam and maintain a Single PMK cache that was set during the
set_pmksa.

Change-Id: Ie803a9816dece79c6d71e9963e11c6ec424306eb
CRs-Fixed: 2798720
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>