qcacld-3.0: Fix to avoid sending invalid PMKID in FT-SuiteB roam
In FT-SuiteB when STA connects to AP1 then PMK1 gets cached. And
then STA disconnects from AP1 and connects to AP2 then PMK2
gets cached. Now if STA roams to AP1 then FW uses PMK2 from mlme
session to create the PMKID. But the corresponding RSO command
from driver sends the PMK1 to FW and the same gets updated in
mlme session of FW. This results in failure of next roams, as the
invalid PMKID gets created using PMK1.
This fix helps in deleting the old/stale PMK cache entries that
are under the same mobility domain as of the newly added entry in
case of all the FT connections except FT-SAE.

Change-Id: Id960e955f8669e19e950c1edc38a9f0a715292b7
CRs-Fixed: 2798713
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>