qcacld-3.0: Fix ll_stats request in progress flag updation
Currently, LL stats in-progress flag is being set first and not getting
reset when peer is invalid. Because of this station stats sent by the
fw are ignored in the host since cp stats request pending is not set.
Also ll_stats request is not being rejected if previous ll_stats
request is in-progress.

To resolve this, update ll_stats request in progress flag after sanity
check for the peer and ignore the new ll_stats request command if the
previous ll_stats request is in progress.

Change-Id: I205502f8c34b51c396859d5ff3ed247e639eea99
CRs-Fixed: 2803032
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>